requirements.txt
flask
python-dotenv
flask-sqlalchemy
flask-login
psycopg2-binary
